backend/requirements.txt
fastapi
uvicorn
python-dotenv
torch
torchaudio
openai-whisper
panns-inference
groq
pydub
requests
